Added HibobConnection implement with custom serializer.  Corrected a few data types in the model.

+++ src/Hibob.Dotnet/Hibob.Dotnet.Core/Model/Employment.cs
@@ -2,7 +2,7 @@
 
 public class Employment
 {
-    public int? Fte { get; set; }
+    public decimal? Fte { get; set; }
     public string? Type { get; set; }
     public object? FlsaCode { get; set; }
     public string? Contract { get; set; }
