+++ src/Hibob.Dotnet/Hibob.Dotnet.Core/Model/Payment.cs
@@ -2,6 +2,6 @@
 
 public class Payment
 {
-    public int? Value { get; set; }
+    public decimal? Value { get; set; }
     public string? Currency { get; set; }
 }
